STA-86: remove Transak dead code and rename package to disbursement
Transak was never wired into any profile and would never be selected
at runtime. Keeping the adapter around doubled the surface area for
the upcoming port-signature change (STA-87) and left a footgun: any
operator who set TRANSAK_ENABLED=true after STA-87 could silently
activate a stale, never-tested code path.

- Delete TransakDisbursementAdapter, TransakConfig, TransakProperties,
  TransakQuoteRequest/Response, TransakOrderRequest/Response, and
  TransakDisbursementAdapterTest.
- Move LoggingDisbursementAdapter, DisbursementConfig, and
  LoggingDisbursementAdapterTest from infrastructure.transak to
  infrastructure.disbursement.
- Remove stablepay.transak.* block from application.yml.
- Replace "Transak API unavailable" strings in temporal test fixtures
  with provider-agnostic wording.

Closes #184
