+++ backend/src/integration-test/java/com/stablepay/infrastructure/temporal/RemittanceLifecycleWorkflowIntegrationTest.java
@@ -187,7 +187,7 @@ class RemittanceLifecycleWorkflowIntegrationTest {
         @Test
         void shouldTransitionToDisbursementFailedWhenDisbursementThrows() {
             // given
-            willThrow(new RuntimeException("Transak API unavailable"))
+            willThrow(new RuntimeException("Disbursement provider unavailable"))
                     .given(activities)
                     .disburseInr(SOME_UPI_ID, SOME_AMOUNT_USDC, remittanceId.toString());
