--- backend/src/main/java/com/stablepay/infrastructure/transak/TransakOrderRequest.java
@@ -1,6 +1,0 @@
-package com.stablepay.infrastructure.transak;
-
-import lombok.Builder;
-
-@Builder(toBuilder = true)
-record TransakOrderRequest(String quoteId, String paymentDetails, String partnerOrderId) {}
