--- backend/src/main/java/com/stablepay/infrastructure/transak/TransakOrderResponse.java
@@ -1,6 +1,0 @@
-package com.stablepay.infrastructure.transak;
-
-import lombok.Builder;
-
-@Builder(toBuilder = true)
-record TransakOrderResponse(String orderId, String status, String depositAddress) {}
