--- backend/src/main/java/com/stablepay/infrastructure/transak/TransakQuoteRequest.java
@@ -1,11 +1,0 @@
-package com.stablepay.infrastructure.transak;
-
-import lombok.Builder;
-
-@Builder(toBuilder = true)
-record TransakQuoteRequest(
-    String cryptoCurrency,
-    String fiatCurrency,
-    String fiatAmount,
-    String type
-) {}
