--- backend/src/main/java/com/stablepay/infrastructure/transak/TransakQuoteResponse.java
@@ -1,6 +1,0 @@
-package com.stablepay.infrastructure.transak;
-
-import lombok.Builder;
-
-@Builder(toBuilder = true)
-record TransakQuoteResponse(String quoteId, String fiatAmount, String cryptoAmount) {}
