+++ backend/src/main/resources/application.yml
@@ -60,11 +60,6 @@ stablepay:
     auto-confirm: ${STRIPE_AUTO_CONFIRM:false}
     test-payment-method: ${STRIPE_TEST_PAYMENT_METHOD:pm_card_visa}
     currency: ${STRIPE_CURRENCY:usd}
-  transak:
-    enabled: ${TRANSAK_ENABLED:false}
-    api-key: ${TRANSAK_API_KEY:}
-    api-secret: ${TRANSAK_API_SECRET:}
-    base-url: ${TRANSAK_BASE_URL:https://staging-api.transak.com}
   twilio:
     enabled: ${TWILIO_ENABLED:false}
     account-sid: ${TWILIO_ACCOUNT_SID:}
